+++ tests/test_hypothesis.py
@@ -164,6 +164,7 @@ def partial_ratio_short_needle(s1, s2):
 
 def cdist_scorer(queries, choices, scorer):
     import numpy as np
+
     matrix = np.zeros((len(queries), len(choices)), dtype=np.uint8)
 
     for i, query in enumerate(queries):
@@ -175,6 +176,7 @@ def cdist_scorer(queries, choices, scorer):
 
 def cdist_distance(queries, choices, scorer):
     import numpy as np
+
     matrix = np.zeros((len(queries), len(choices)), dtype=np.int32)
 
     for i, query in enumerate(queries):